# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/participantes/{id}": {
      "get": {
        "operationId": "ParticipanteController_obtenerPorId",
        "summary": "API para obtener un participante por su ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del participante que deseas obtener",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Participante"
        ]
      },
      "patch": {
        "operationId": "ParticipanteController_actualizar",
        "summary": "API para actualizar un participante",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del participante que deseas modificar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Esta API permite actualizar un participante existente utilizando los atributos proporcionados en el cuerpo de la solicitud.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarParticipanteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Participante"
        ]
      }
    },
    "/participantes/eliminar/{id}": {
      "delete": {
        "operationId": "ParticipanteController_eliminar",
        "summary": "API para eliminar un participante por su ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del participante que deseas eliminar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Participante"
        ]
      }
    },
    "/participantes/soft-delete/{id}": {
      "patch": {
        "operationId": "ParticipanteController_softDelete",
        "summary": "API para eliminar un participante por su ID (soft delete)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del participante que deseas eliminar (soft delete)",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Participante"
        ]
      }
    },
    "/participantes/restore/{id}": {
      "patch": {
        "operationId": "ParticipanteController_restore",
        "summary": "API para restaurar un participante eliminado (soft delete)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del participante que deseas restaurar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Participante"
        ]
      }
    },
    "/equipos": {
      "get": {
        "operationId": "EquipoController_listar",
        "summary": "API para obtener el listado de equipos",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Equipo"
        ]
      },
      "post": {
        "operationId": "EquipoController_crear",
        "summary": "API para crear un nuevo equipo",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Esta API permite crear un nuevo equipo utilizando los datos proporcionados en el cuerpo de la solicitud.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearEquipoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Equipo"
        ]
      }
    },
...
    "/equipos/{id}": {
      "get": {
        "operationId": "EquipoController_obtenerPorId",
        "summary": "API para obtener un equipo por su ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del equipo que deseas obtener",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Equipo"
        ]
      },
      "patch": {
        "operationId": "EquipoController_actualizar",
        "summary": "API para actualizar un equipo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del equipo que deseas modificar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Esta API permite actualizar un equipo existente utilizando los atributos proporcionados en el cuerpo de la solicitud.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarEquipoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Equipo"
        ]
      }
    },
    "/equipos/eliminar/{id}": {
      "delete": {
        "operationId": "EquipoController_eliminar",
        "summary": "API para eliminar un equipo por su ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del equipo que deseas eliminar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Equipo"
        ]
      }
    },
    "/equipos/soft-delete/{id}": {
      "patch": {
        "operationId": "EquipoController_softDelete",
        "summary": "API para eliminar un equipo por su ID (soft delete)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del equipo que deseas eliminar (soft delete)",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Equipo"
        ]
      }
    },
    "/equipos/restore/{id}": {
      "patch": {
        "operationId": "EquipoController_restore",
        "summary": "API para restaurar un equipo eliminado (soft delete)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del equipo que deseas restaurar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Equipo"
        ]
      }
    },
    "/equiposParticipantes": {
      "get": {
        "operationId": "EquipoParticipanteController_listar",
        "summary": "API para obtener el listado de equiposParticipantes",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EquipoParticipante"
        ]
      },
      "post": {
        "operationId": "EquipoParticipanteController_crear",
        "summary": "API para crear un nuevo equipoParticipante",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Esta API permite crear un nuevo equipoParticipante utilizando los datos proporcionados en el cuerpo de la solicitud.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearEquipoParticipanteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "EquipoParticipante"
        ]
      }
    },
    "/equiposParticipantes/{id}": {
      "get": {
        "operationId": "EquipoParticipanteController_obtenerPorId",
        "summary": "API para obtener un equipoParticipante por su ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del equipoParticipante que deseas obtener",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EquipoParticipante"
        ]
      },
      "patch": {
        "operationId": "EquipoParticipanteController_actualizar",
        "summary": "API para actualizar un equipoParticipante",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del equipoParticipante que deseas modificar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Esta API permite actualizar un equipoParticipante existente utilizando los atributos proporcionados en el cuerpo de la solicitud.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarEquipoParticipanteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EquipoParticipante"
        ]
      }
    },
    "/equiposParticipantes/eliminar/{id}": {
      "delete": {
        "operationId": "EquipoParticipanteController_eliminar",
        "summary": "API para eliminar un equipoParticipante por su ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del equipoParticipante que deseas eliminar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EquipoParticipante"
        ]
      }
    },
    "/equiposParticipantes/soft-delete/{id}": {
      "patch": {
        "operationId": "EquipoParticipanteController_softDelete",
        "summary": "API para eliminar un equipoParticipante por su ID (soft delete)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del equipoParticipante que deseas eliminar (soft delete)",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EquipoParticipante"
        ]
      }
    },
    "/equiposParticipantes/restore/{id}": {
      "patch": {
        "operationId": "EquipoParticipanteController_restore",
        "summary": "API para restaurar un equipoParticipante eliminado (soft delete)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del equipoParticipante que deseas restaurar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EquipoParticipante"
        ]
      }
    },
...
  },
  "info": {
    "title": "Backend",
    "description": "Documentación",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
...
      "ActualizarParticipanteDto": {
        "type": "object",
        "properties": {
          "nombreCompleto": {
            "type": "string",
            "example": "Arturo Rodrigo Perez Chavez"
          },
          "carnetIdentidad": {
            "type": "string",
            "example": "8526947"
          },
          "fechaNacimiento": {
            "type": "string",
            "example": "2003-05-23"
          },
          "departamento": {
            "type": "string",
            "example": "La Paz"
          },
          "provincia": {
            "type": "string",
            "example": "Murillo"
          },
          "municipio": {
            "type": "string",
            "example": "Achocalla"
          }
        },
        "required": [
          "nombreCompleto",
          "carnetIdentidad",
          "fechaNacimiento",
          "departamento",
          "provincia",
          "municipio"
        ]
      },
...
      "CrearEquipoDto": {
        "type": "object",
        "properties": {
          "nombreEquipo": {
            "type": "string",
            "example": "Equipo 1"
          }
        },
        "required": [
          "nombreEquipo"
        ]
      },
      "ActualizarEquipoDto": {
        "type": "object",
        "properties": {
          "nombreEquipo": {
            "type": "string",
            "example": "Equipo 1"
          }
        },
        "required": [
          "nombreEquipo"
        ]
      },
      "CrearEquipoParticipanteDto": {
        "type": "object",
        "properties": {
          "equipoId": {
            "type": "number",
            "description": "ID del equipo (foreign key)",
            "format": "int64",
            "example": "1"
          },
          "participanteId": {
            "type": "number",
            "description": "ID del participante (foreign key)",
            "format": "int64",
            "example": "1"
          }
        },
        "required": [
          "equipoId",
          "participanteId"
        ]
      },
      "ActualizarEquipoParticipanteDto": {
        "type": "object",
        "properties": {
          "equipo": {
            "type": "number",
            "description": "ID del equipo (foreign key)",
            "format": "int64",
            "example": "1"
          },
          "participante": {
            "type": "number",
            "description": "ID del participante (foreign key)",
            "format": "int64",
            "example": "1"
          }
        },
        "required": [
          "equipo",
          "participante"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}